+++ src/solvers/dg_2d.jl
@@ -1124,8 +1124,8 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{2}, nonconservative_terms::
     # For `Float32`, this gives 1.1920929f-5
     atol = 1.8189894035458565e-12 # see also `pure_and_blended_element_ids!` in Trixi.jl
 
-    element_ids_dg = zero(CuArray{Int64}(undef, length(alpha)))
-    element_ids_dgfv = zero(CuArray{Int64}(undef, length(alpha)))
+    element_ids_dg = zero(CuArray{Int}(undef, length(alpha)))
+    element_ids_dgfv = zero(CuArray{Int}(undef, length(alpha)))
 
     pure_blended_element_count_kernel = @cuda launch=false pure_blended_element_count_kernel!(element_ids_dg,
                                                                                               element_ids_dgfv,
@@ -1207,8 +1207,8 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{2}, nonconservative_terms::
     # For `Float32`, this gives 1.1920929f-5
     atol = 1.8189894035458565e-12 # see also `pure_and_blended_element_ids!` in Trixi.jl
 
-    element_ids_dg = zero(CuArray{Int64}(undef, length(alpha)))
-    element_ids_dgfv = zero(CuArray{Int64}(undef, length(alpha)))
+    element_ids_dg = zero(CuArray{Int}(undef, length(alpha)))
+    element_ids_dgfv = zero(CuArray{Int}(undef, length(alpha)))
 
     pure_blended_element_count_kernel = @cuda launch=false pure_blended_element_count_kernel!(element_ids_dg,
                                                                                               element_ids_dgfv,
@@ -1292,8 +1292,8 @@ end
 
 # Pack kernels for prolonging solution to interfaces
 function cuda_prolong2interfaces!(u, mesh::TreeMesh{2}, equations, cache)
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
 
     size_arr = CuArray{Float64}(undef, size(interfaces_u, 2) * size(interfaces_u, 3),
@@ -1316,8 +1316,8 @@ function cuda_interface_flux!(mesh::TreeMesh{2}, nonconservative_terms::False, e
                               cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
     surface_flux_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
     surface_flux_values = CuArray{Float64}(cache.elements.surface_flux_values)
@@ -1351,8 +1351,8 @@ function cuda_interface_flux!(mesh::TreeMesh{2}, nonconservative_terms::True, eq
                               cache)
     surface_flux, nonconservative_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
     surface_flux_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
     noncons_left_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
@@ -1400,9 +1400,9 @@ end
 # Pack kernels for prolonging solution to boundaries
 function cuda_prolong2boundaries!(u, mesh::TreeMesh{2}, boundary_conditions::NamedTuple, equations,
                                   cache)
-    neighbor_ids = CuArray{Int64}(cache.boundaries.neighbor_ids)
-    neighbor_sides = CuArray{Int64}(cache.boundaries.neighbor_sides)
-    orientations = CuArray{Int64}(cache.boundaries.orientations)
+    neighbor_ids = CuArray{Int}(cache.boundaries.neighbor_ids)
+    neighbor_sides = CuArray{Int}(cache.boundaries.neighbor_sides)
+    orientations = CuArray{Int}(cache.boundaries.orientations)
     boundaries_u = CuArray{Float64}(cache.boundaries.u)
 
     size_arr = CuArray{Float64}(undef, size(boundaries_u, 2) * size(boundaries_u, 3),
@@ -1433,10 +1433,10 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{2}, boundary_conditions::NamedTup
                              nonconservative_terms::False, equations, dg::DGSEM, cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    n_boundaries_per_direction = CuArray{Int64}(cache.boundaries.n_boundaries_per_direction)
-    neighbor_ids = CuArray{Int64}(cache.boundaries.neighbor_ids)
-    neighbor_sides = CuArray{Int64}(cache.boundaries.neighbor_sides)
-    orientations = CuArray{Int64}(cache.boundaries.orientations)
+    n_boundaries_per_direction = CuArray{Int}(cache.boundaries.n_boundaries_per_direction)
+    neighbor_ids = CuArray{Int}(cache.boundaries.neighbor_ids)
+    neighbor_sides = CuArray{Int}(cache.boundaries.neighbor_sides)
+    orientations = CuArray{Int}(cache.boundaries.orientations)
     boundaries_u = CuArray{Float64}(cache.boundaries.u)
     node_coordinates = CuArray{Float64}(cache.boundaries.node_coordinates)
     surface_flux_values = CuArray{Float64}(cache.elements.surface_flux_values)
@@ -1450,8 +1450,8 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{2}, boundary_conditions::NamedTup
                               configurator_1d(last_first_indices_kernel, lasts)...)
 
     lasts, firsts = Array(lasts), Array(firsts)
-    boundary_arr = CuArray{Int64}(firsts[1]:lasts[4])
-    indices_arr = CuArray{Int64}([firsts[1], firsts[2], firsts[3], firsts[4]])
+    boundary_arr = CuArray{Int}(firsts[1]:lasts[4])
+    indices_arr = CuArray{Int}([firsts[1], firsts[2], firsts[3], firsts[4]])
     boundary_conditions_callable = replace_boundary_conditions(boundary_conditions)
 
     size_arr = CuArray{Float64}(undef, size(surface_flux_values, 2), length(boundary_arr))
@@ -1479,10 +1479,10 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{2}, boundary_conditions::NamedTup
                              nonconservative_terms::True, equations, dg::DGSEM, cache)
     surface_flux, nonconservative_flux = dg.surface_integral.surface_flux
 
-    n_boundaries_per_direction = CuArray{Int64}(cache.boundaries.n_boundaries_per_direction)
-    neighbor_ids = CuArray{Int64}(cache.boundaries.neighbor_ids)
-    neighbor_sides = CuArray{Int64}(cache.boundaries.neighbor_sides)
-    orientations = CuArray{Int64}(cache.boundaries.orientations)
+    n_boundaries_per_direction = CuArray{Int}(cache.boundaries.n_boundaries_per_direction)
+    neighbor_ids = CuArray{Int}(cache.boundaries.neighbor_ids)
+    neighbor_sides = CuArray{Int}(cache.boundaries.neighbor_sides)
+    orientations = CuArray{Int}(cache.boundaries.orientations)
     boundaries_u = CuArray{Float64}(cache.boundaries.u)
     node_coordinates = CuArray{Float64}(cache.boundaries.node_coordinates)
     surface_flux_values = CuArray{Float64}(cache.elements.surface_flux_values)
@@ -1496,8 +1496,8 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{2}, boundary_conditions::NamedTup
                               configurator_1d(last_first_indices_kernel, lasts)...)
 
     lasts, firsts = Array(lasts), Array(firsts)
-    boundary_arr = CuArray{Int64}(firsts[1]:lasts[4])
-    indices_arr = CuArray{Int64}([firsts[1], firsts[2], firsts[3], firsts[4]])
+    boundary_arr = CuArray{Int}(firsts[1]:lasts[4])
+    indices_arr = CuArray{Int}([firsts[1], firsts[2], firsts[3], firsts[4]])
 
     # Replace with callable functions (not necessary here)
     # boundary_conditions_callable = replace_boundary_conditions(boundary_conditions)
@@ -1530,9 +1530,9 @@ end
 
 # Pack kernels for prolonging solution to mortars
 function cuda_prolong2mortars!(u, mesh::TreeMesh{2}, cache_mortars::True, dg::DGSEM, cache)
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper = zero(CuArray{Float64}(cache.mortars.u_upper)) # NaN to zero
     u_lower = zero(CuArray{Float64}(cache.mortars.u_lower)) # NaN to zero
@@ -1582,9 +1582,9 @@ function cuda_mortar_flux!(mesh::TreeMesh{2}, cache_mortars::True, nonconservati
                            equations, dg::DGSEM, cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper = CuArray{Float64}(cache.mortars.u_upper)
     u_lower = CuArray{Float64}(cache.mortars.u_lower)
@@ -1633,9 +1633,9 @@ function cuda_mortar_flux!(mesh::TreeMesh{2}, cache_mortars::True, nonconservati
                            equations, dg::DGSEM, cache)
     surface_flux, nonconservative_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper = CuArray{Float64}(cache.mortars.u_upper)
     u_lower = CuArray{Float64}(cache.mortars.u_lower)
