+++ src/solvers/dg_3d.jl
@@ -1450,8 +1450,8 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{3}, nonconservative_terms::
     # For `Float32`, this gives 1.1920929f-5
     atol = 1.8189894035458565e-12 # see also `pure_and_blended_element_ids!` in Trixi.jl
 
-    element_ids_dg = zero(CuArray{Int64}(undef, length(alpha)))
-    element_ids_dgfv = zero(CuArray{Int64}(undef, length(alpha)))
+    element_ids_dg = zero(CuArray{Int}(undef, length(alpha)))
+    element_ids_dgfv = zero(CuArray{Int}(undef, length(alpha)))
 
     pure_blended_element_count_kernel = @cuda launch=false pure_blended_element_count_kernel!(element_ids_dg,
                                                                                               element_ids_dgfv,
@@ -1542,8 +1542,8 @@ function cuda_volume_integral!(du, u, mesh::TreeMesh{3}, nonconservative_terms::
     # For `Float32`, this gives 1.1920929f-5
     atol = 1.8189894035458565e-12 # see also `pure_and_blended_element_ids!` in Trixi.jl
 
-    element_ids_dg = zero(CuArray{Int64}(undef, length(alpha)))
-    element_ids_dgfv = zero(CuArray{Int64}(undef, length(alpha)))
+    element_ids_dg = zero(CuArray{Int}(undef, length(alpha)))
+    element_ids_dgfv = zero(CuArray{Int}(undef, length(alpha)))
 
     pure_blended_element_count_kernel = @cuda launch=false pure_blended_element_count_kernel!(element_ids_dg,
                                                                                               element_ids_dgfv,
@@ -1640,8 +1640,8 @@ end
 
 # Pack kernels to prolonging solution to interfaces
 function cuda_prolong2interfaces!(u, mesh::TreeMesh{3}, equations, cache)
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
 
     size_arr = CuArray{Float64}(undef, size(interfaces_u, 2) * size(interfaces_u, 3)^2,
@@ -1664,8 +1664,8 @@ function cuda_interface_flux!(mesh::TreeMesh{3}, nonconservative_terms::False, e
                               cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
     surface_flux_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
     surface_flux_values = CuArray{Float64}(cache.elements.surface_flux_values)
@@ -1764,8 +1764,8 @@ function cuda_interface_flux!(mesh::TreeMesh{3}, nonconservative_terms::True, eq
                               cache)
     surface_flux, nonconservative_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.interfaces.neighbor_ids)
-    orientations = CuArray{Int64}(cache.interfaces.orientations)
+    neighbor_ids = CuArray{Int}(cache.interfaces.neighbor_ids)
+    orientations = CuArray{Int}(cache.interfaces.orientations)
     interfaces_u = CuArray{Float64}(cache.interfaces.u)
     surface_flux_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
     noncons_left_arr = CuArray{Float64}(undef, size(interfaces_u)[2:end]...)
@@ -1815,9 +1815,9 @@ end
 # Pack kernels for prolonging solution to boundaries
 function cuda_prolong2boundaries!(u, mesh::TreeMesh{3}, boundary_conditions::NamedTuple, equations,
                                   cache)
-    neighbor_ids = CuArray{Int64}(cache.boundaries.neighbor_ids)
-    neighbor_sides = CuArray{Int64}(cache.boundaries.neighbor_sides)
-    orientations = CuArray{Int64}(cache.boundaries.orientations)
+    neighbor_ids = CuArray{Int}(cache.boundaries.neighbor_ids)
+    neighbor_sides = CuArray{Int}(cache.boundaries.neighbor_sides)
+    orientations = CuArray{Int}(cache.boundaries.orientations)
     boundaries_u = CuArray{Float64}(cache.boundaries.u)
 
     size_arr = CuArray{Float64}(undef, size(boundaries_u, 2) * size(boundaries_u, 3)^2,
@@ -1848,10 +1848,10 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{3}, boundary_conditions::NamedTup
                              nonconservative_terms, equations, dg::DGSEM, cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    n_boundaries_per_direction = CuArray{Int64}(cache.boundaries.n_boundaries_per_direction)
-    neighbor_ids = CuArray{Int64}(cache.boundaries.neighbor_ids)
-    neighbor_sides = CuArray{Int64}(cache.boundaries.neighbor_sides)
-    orientations = CuArray{Int64}(cache.boundaries.orientations)
+    n_boundaries_per_direction = CuArray{Int}(cache.boundaries.n_boundaries_per_direction)
+    neighbor_ids = CuArray{Int}(cache.boundaries.neighbor_ids)
+    neighbor_sides = CuArray{Int}(cache.boundaries.neighbor_sides)
+    orientations = CuArray{Int}(cache.boundaries.orientations)
     boundaries_u = CuArray{Float64}(cache.boundaries.u)
     node_coordinates = CuArray{Float64}(cache.boundaries.node_coordinates)
     surface_flux_values = CuArray{Float64}(cache.elements.surface_flux_values)
@@ -1865,8 +1865,8 @@ function cuda_boundary_flux!(t, mesh::TreeMesh{3}, boundary_conditions::NamedTup
                               configurator_1d(last_first_indices_kernel, lasts)...)
 
     lasts, firsts = Array(lasts), Array(firsts)
-    boundary_arr = CuArray{Int64}(firsts[1]:lasts[6])
-    indices_arr = CuArray{Int64}([firsts[1], firsts[2], firsts[3], firsts[4], firsts[5], firsts[6]])
+    boundary_arr = CuArray{Int}(firsts[1]:lasts[6])
+    indices_arr = CuArray{Int}([firsts[1], firsts[2], firsts[3], firsts[4], firsts[5], firsts[6]])
     boundary_conditions_callable = replace_boundary_conditions(boundary_conditions)
 
     size_arr = CuArray{Float64}(undef, size(surface_flux_values, 2)^2, length(boundary_arr))
@@ -1895,9 +1895,9 @@ end
 
 # Pack kernels for prolonging solution to mortars
 function cuda_prolong2mortars!(u, mesh::TreeMesh{3}, cache_mortars::True, dg::DGSEM, cache)
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper_left = zero(CuArray{Float64}(cache.mortars.u_upper_left)) # NaN to zero
     u_upper_right = zero(CuArray{Float64}(cache.mortars.u_upper_right)) # NaN to zero
@@ -1981,9 +1981,9 @@ function cuda_mortar_flux!(mesh::TreeMesh{3}, cache_mortars::True, nonconservati
                            equations, dg::DGSEM, cache)
     surface_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper_left = CuArray{Float64}(cache.mortars.u_upper_left)
     u_upper_right = CuArray{Float64}(cache.mortars.u_upper_right)
@@ -2069,9 +2069,9 @@ function cuda_mortar_flux!(mesh::TreeMesh{3}, cache_mortars::True, nonconservati
                            equations, dg::DGSEM, cache)
     surface_flux, nonconservative_flux = dg.surface_integral.surface_flux
 
-    neighbor_ids = CuArray{Int64}(cache.mortars.neighbor_ids)
-    large_sides = CuArray{Int64}(cache.mortars.large_sides)
-    orientations = CuArray{Int64}(cache.mortars.orientations)
+    neighbor_ids = CuArray{Int}(cache.mortars.neighbor_ids)
+    large_sides = CuArray{Int}(cache.mortars.large_sides)
+    orientations = CuArray{Int}(cache.mortars.orientations)
 
     u_upper_left = CuArray{Float64}(cache.mortars.u_upper_left)
     u_upper_right = CuArray{Float64}(cache.mortars.u_upper_right)
